+++ CHANGES.rst
@@ -9,6 +9,8 @@ WIP
   Added Python3.14 support.
   Added type annotations and ``py.typed`` file.
   Now ``Container.child`` returns ``Self`` type, not ``Container`` type.
+  Changes how ``typing.ForwardRef`` instances are compared: first we compare them
+  by hash, next by ``__forward_arg__``.
 
 `0.7.0`_ 2023-09-11
 -------------------